Removed modifications from existing migrations

+++ database/migrations/2023_02_05_224726_create_statistics_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('statistics', function (Blueprint $table) {
             $table->id();
             $table->integer('property_id');
-            $table->decimal('price',14,0);
+            $table->integer('price');
             $table->timestamps();
         });
     }
